20240318 스프링

+++ demo/src/main/java/com/example/demo/emp/SearchVO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class SearchVO {
 	
-	 int start = 1;
+	long start = 1;
 	 int end = 10;
 	 
 	 int[] employeeIds;
